4.1
добавил абстрактный класс и наследовал от него свойства для всех компонент
