+++ src/view/empty-points-list.js
@@ -1,23 +1,9 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createEmptyList = () => '<p class="trip-events__msg">Click New Event to create your first point</p>';
 
-export default class MessageWithoutPoints {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class MessageWithoutPoints extends AbstractView{
   get template() {
     return createEmptyList();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
