+++ src/view/filter-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 
 const createFilterTemplate = (
@@ -21,22 +21,8 @@ const createFilterTemplate = (
     <button class="visually-hidden" type="submit">Accept filter</button>
   </form> `);
 
-export default class FilterView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class FilterView extends AbstractView {
   get template() {
     return createFilterTemplate;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
