+++ src/view/point-list-view.js
@@ -1,23 +1,9 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createPointListTemplate = () => '<div class="trip-events__list"></div>';
 
-export default class PointListView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class PointListView extends AbstractView {
   get template() {
     return createPointListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
