+++ src/view/site-menu-view.js
@@ -1,5 +1,4 @@
-import {createElement} from '../render.js';
-
+import AbstractView from './abstract-view.js';
 
 const createSiteMenuTemplate = () =>
   `<div class="trip-controls__navigation">
@@ -11,22 +10,8 @@ const createSiteMenuTemplate = () =>
   </div>
   `;
 
-export default class SiteMenuView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class SiteMenuView extends AbstractView{
   get template() {
     return createSiteMenuTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
