+++ src/view/sort-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const CreateSortTemplate = (
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
@@ -30,22 +30,8 @@ const CreateSortTemplate = (
   `);
 
 
-export default class SortView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class SortView extends AbstractView{
   get template() {
     return CreateSortTemplate;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
